Fix integrity propagation bug
Deriving primitive classes from each does not work as anticipated.
The constructor of the direct base class did not get called,
instead the one of the class Primitive. This lead to missing rules.

I created new instances instead of deriving for now. While this
works, some rules got duplicated which can be a maintenance issue.

FIXME: I need to revise this. #1